pyproject.toml
[project]
name = "thai-plate-ocr"
version = "0.1.0"
description = "Two-stage Thai license-plate OCR: YOLOv8 plate detector + YOLOv8 character detector with spatial post-processing for single- and two-line Thai plates."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Simon Yosboon" }]
license = { text = "MIT" }
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "pillow>=10.4",
    "opencv-python-headless>=4.10",
    "ultralytics>=8.3",
    "torch>=2.4",
    "torchvision>=0.19",
    "roboflow>=1.1",
    "matplotlib>=3.9",
    "mlflow>=2.17",
    "typer>=0.12",
    "rich>=13.9",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.12",
    "pydantic>=2.9",
    "tqdm>=4.66",
    "tabulate>=0.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "httpx>=0.27",
]

[project.scripts]
plate = "thai_plate_ocr.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/thai_plate_ocr"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
